requirements.txt
streamlit
python-dotenv

langchain
langchain-community
langchain-google-genai
langchain-text-splitters

google-generativeai

chromadb
sentence-transformers
pypdf